修复类型不一致问题代码更新V3.6

+++ data-middle/src/main/scala/com/yzs/data/common/clickHouseInsert.scala
@@ -89,7 +89,7 @@ class clickHouseInsert extends Serializable {
         //     val Ck = println(11111)
         dataTypeTemp = "Float"
         dataTemp = dealDefaultValue(columnValueTemp, dataTypeTemp)
-        prepareState.setDouble(i, dataTemp.toString.toInt)
+        prepareState.setDouble(i, dataTemp.toString.toFloat)
       }
       case _ => {
         prepareState.setString(i, columnValueTemp)
